Finalizado geração de arquivos de token, tabela e erros.

+++ Codigo/Header/tokens.h
@@ -87,7 +87,7 @@ typedef struct
 typedef struct
 {
     tTokenEncontrado *tab;    
-    int tamAtual;
+    size_t tamAtual;
     size_t tamLimite;
 } tTabTokenEncontrados;
